requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cfgv==3.4.0
chardet==5.2.0
click==8.1.8
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
flexcache==0.3
flexparser==0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-docx==1.1.2
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
virtualenv==20.30.0
